+++ changelog.md
@@ -4,7 +4,7 @@
 * splitByTag now works for clients
 * fix a rare case where a `.DeepClone()` was appended to a dictionary of additional properties.
 * default is nullable now sets add `?` at the end of value objects if `nullableIsDefault` is set to true.
-
+* if content type */* is found, application/json is assumed
 ## 4.3.0
 * new configuration option: `nullableIsDefault`. If set to true, all non-required properties with no will be treated as nullable, and the default value will be null if no default is defined. the option can be overriden either way by setting `x-caffoa-nullable` to true or false on an schema item.